Achievement Update
업적 업데이트 단일 함수에서 수행하게끔 변경

+++ Source/ProjectDream/GameSystems/Inventory/DreamItemDTO.h
@@ -29,7 +29,7 @@ public:
 	bool operator !=(const FDreamGameItemDef& Other) const;
 	int32 GetItemID() const { return ItemID; }
 	int32 GetMaxStackCnt() const { return MaxStackCnt; }
-	int32 GetItemWeight() const { return ItemWeight; }
+	float GetItemWeight() const { return ItemWeight; }
 public:
 	UPROPERTY(EditAnywhere, Category = "Data")		 EItemCategory ItemCategory;
 	UPROPERTY(EditAnywhere, Category = "Data")		 FString ItemName;
